Data
The files of "data_n*_m1000_sam100" are the data of optimal strategies under varying organismal size n. Here the "data_n5_m1000_sam100" are missed as the data size is out the maximum size for uploading. It is produced in the same way as others by using the file "dynamic_heatmap.py" in code part by setting the variable "num_duplicate" equals to 100. “data_bc12” is used to produce the figure4BCD in the main text.